@@ -38,9 +38,9 @@ from .exceptions import (
     InvalidActionError,
 )
 
-__version__ = "1.0.0"
-__author__ = "Your Name"
-__email__ = "your.email@example.com"
+__version__ = "0.1.0"
+__author__ = "Eowl"
+__email__ = "eowl@me.com"
 
 __all__ = [
     "ActionDispatcher",
